requirements.txt
fastapi>=0.111,<0.116
uvicorn[standard]>=0.30,<0.33

# Environment variables
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0

# Google Gmail API
google-auth>=2.23.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth-httplib2>=0.1.1,<0.2.0
google-api-python-client>=2.100.0,<3.0.0
google-cloud-pubsub>=2.21.0,<3.0.0

# PDF Processing
pdfminer.six>=20221105,<20240000
PyPDF2>=3.0.0,<4.0.0
